\name{pollackCN16}
\alias{pollackCN16}
\docType{data}
\title{ Breast cancer data (copy number) }
\description{
Copy number data of chromosome 16 the breast cancer data set. Called using \code{CGHcall} with default settings,
contains 240 features and 41 samples.
}
\usage{data(pollackCN16)}
\format{
An object of class \code{\link{cghCall}}.
}
\source{
Pollack, J. R., Sorlie, T., Perou, C. M., Rees, C. A., Jeffrey, S. S., 
Lonning, P. E., Tibshirani, R., Botstein, D., Borresen-
Dale, A. L., Brown, P. O. (2002), "Microarray analysis 
reveals a major direct role of DNA copy number alteration
in the transcriptional program of human breast tumors", \emph{PNAS}, 99, 12963-12968.
}
\references{
Van de Wiel, M. A., Kim, K. I., Vosse, S. J., Van Wieringen, W. N., Wilting, S. M., Ylstra, B. (2007), 
"CGHcall: Calling aberrations for array CGH tumor profiles", \emph{Bioinformatics}, 23, 892-894.
}
\examples{
data(pollackCN16)
}
\keyword{datasets}
