\name{pollackGE16}
\alias{pollackGE16}
\docType{data}
\title{ Breast cancer data (gene expression) }
\description{
Gene expression data of chromosome 16 of the breast cancer data set; contains 240 features and 41 samples.
}
\usage{data(pollackGE16)}
\format{
An object of class \code{\link{ExpressionSet}}.
}
\source{
Pollack, J.R., Sorlie, T., Perou, C.M., Rees, C.A., Jeffrey, S.S., 
Lonning, P.E., Tibshirani, R., Botstein, D., Borresen-
Dale, A.L., Brown, P.O. (2002), "Microarray analysis 
reveals a major direct role of DNA copy number alteration
in the transcriptional program of human breast tumors", 
\emph{PNAS}, 99, 12963-12968.
}
\examples{
data(pollackGE16)
}
\keyword{datasets}
