\name{RCMrandom-method}
\docType{methods}
\alias{RCMrandom-method}
\alias{RCMrandom,rcmFit-method}
\title{Methods for Function RCMrandom}
\description{
Methods for function \code{RCMrandom} 
}
\section{Methods}{
\describe{
\item{\code{signature(object = "rcmFit")}}{
Draws random data of same dimension as data on which the \code{rcmFit}-object was fitted.
}
}
}
\keyword{methods}

