\name{sigaR-package}
\alias{sigaR-package}
\alias{sigaR}
\docType{package}
\title{
statistics for integrative genomics analyses in R}
\description{
The package facilitates several types of integrative analysis of high-throughput data from various molecular levels.
In addition, it includes functions for data management and visualization.
}
\details{
\tabular{ll}{
Package: \tab sigaR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-04-15\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Wessel N. van Wieringen
Maintainer: Wessel N. van Wieringen <w.vanwieringen@vumc.nl>
}
\references{
Van Wieringen, W.N., Van de Wiel, M.A. (2009), "Non-parametric testing for DNA copy number induced differential mRNA gene expression", \emph{Biometrics}, 65(1), 19-29. 

Van Wieringen, W.N., Berkhof, J., Van de Wiel, M.A. (2010), "A random coefficients model for regional co-expression associated with DNA copy number", \emph{Statistical Applications in Genetics and Molecular Biology}, Volume 9, Issue1, Article 25, 1-28. 

Van Wieringen, W.N., Van der Vaart, A.W. (2011), "Statistical analysis of the cancer cell's molecular entropy using high-throughput data", \emph{Bioinformatics}, 27(4), 556-563. 

Van Wieringen, W.N., Unger, K., Leday, G.G.R., Krijgsman, O., De Menezes, R.X., Ylstra, B., Van de Wiel, M.A. (2011), "Matching of array CGH and gene expression microarray features for the purpose of integrative analysis", \emph{BMC Bioinformatics}, 13:80. 

Van Wieringen, W.N., Van de Wiel, M.A. (2012), "Modeling the \emph{cis}- and \emph{trans}-effect of DNA copy number aberrations on gene expression levels in a pathway", \emph{submitted for publication}. 
}
\keyword{ package }
