\name{summary-method}
\docType{methods}
\alias{summary-method}
\alias{summary,ANY-method}
\alias{summary,entTest-method}
\alias{summary,miTest-method}
\alias{summary,rcmFit-method}
\alias{summary,rcmTest-method}
\title{ Methods for Function summary }
\description{
Methods for function \code{summary}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ANY")}}{ Regular. }

\item{\code{signature(object = "entTest")}}{ Print output. }

\item{\code{signature(object = "miTest")}}{ Print output. }

\item{\code{signature(object = "rcmFit")}}{ Print output. }

\item{\code{signature(object = "rcmTest")}}{ Print output. }
}}
\keyword{methods}

